mysite/requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
